[rename] 테스트 이름 기능 중심으로 변경 repository -> saveFavoriteEvent (#31)
