+++ src/main/java/hyangyu/server/saveFavoriteEvent/UserRepository.java
@@ -1,4 +1,4 @@
-package hyangyu.server.repository;
+package hyangyu.server.saveFavoriteEvent;
 
 import java.util.Optional;
