+++ src/main/java/hyangyu/server/service/FairService.java
@@ -1,9 +1,7 @@
 package hyangyu.server.service;
 
-import hyangyu.server.domain.Display;
 import hyangyu.server.domain.Fair;
-import hyangyu.server.repository.DisplayRepository;
-import hyangyu.server.repository.FairRepository;
+import hyangyu.server.saveFavoriteEvent.FairRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
