+++ src/main/java/hyangyu/server/service/FavoriteDisplayService.java
@@ -3,9 +3,9 @@ package hyangyu.server.service;
 import hyangyu.server.domain.Display;
 import hyangyu.server.domain.FavoriteDisplay;
 import hyangyu.server.domain.User;
-import hyangyu.server.repository.DisplayRepository;
-import hyangyu.server.repository.FavoriteDisplayRepository;
-import hyangyu.server.repository.UserRepository;
+import hyangyu.server.saveFavoriteEvent.DisplayRepository;
+import hyangyu.server.saveFavoriteEvent.FavoriteDisplayRepository;
+import hyangyu.server.saveFavoriteEvent.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
