+++ src/test/java/hyangyu/server/saveFavoriteEvent/FavoriteDisplayRepositoryTest.java
@@ -1,4 +1,4 @@
-package hyangyu.server.repository;
+package hyangyu.server.saveFavoriteEvent;
 
 import hyangyu.server.domain.Display;
 import hyangyu.server.domain.FavoriteDisplay;
